backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Payment Processing
stripe==7.12.0

# HTTP Client (for catalogue parsing)
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Web Scraping
firecrawl-py>=2.0.0

# Scheduling
apscheduler==3.10.4

# Caching
redis==5.0.1

# Email
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Image Processing
Pillow==10.2.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
